arcade-agent/requirements.txt
Flask
psutil
requests
Pillow
packaging
google-genai
